Fix title positioning and split out slider creation
